+++ .dockerignore
@@ -7,3 +7,4 @@ README.md
 .buildHelper.txt
 screenshots/*
 docs/*
+.git
