+++ app/templates/requests.html
@@ -13,9 +13,7 @@ Download Plex Client
         {{ _("We are excited to offer you a wide selection of media to choose from. If you're having trouble finding something you like, don't worry! We have a user-friendly request system that can automatically search for the media you're looking for. ") }}
       </p>
     </div>
-
   </div>
-
 </div>
 <div class="flex flex-col items-center justify-center px-6 py-2 mx-auto lg:py-0">
   <div class="p-6 max-w-sm bg-white rounded-lg border border-gray-200 shadow-md dark:bg-gray-800 dark:border-gray-700">
